Retarget to Revit 2025 (.NET 8, updated APIs)
- .csproj: net48 → net8.0-windows, Revit 2025 API paths, EnableDynamicLoading
- ElementId.IntegerValue → ElementId.Value (long) throughout
- new ElementId(int) → new ElementId(long) for Revit 2025
- ParameterGroup enum → ForgeTypeId via GetGroupTypeId() / GetLabelForGroup()
- ElementData.RevitId/ExpressId changed from int to long
- Removed System.Net.Http reference (built into .NET 8)

+++ ClashControlConnector/Protocol/ElementData.cs
@@ -6,12 +6,12 @@ namespace ClashControlConnector.Protocol
     public class ElementData
     {
         [JsonProperty("globalId")] public string GlobalId { get; set; }
-        [JsonProperty("expressId")] public int ExpressId { get; set; }
+        [JsonProperty("expressId")] public long ExpressId { get; set; }
         [JsonProperty("category")] public string Category { get; set; }
         [JsonProperty("name")] public string Name { get; set; }
         [JsonProperty("level")] public string Level { get; set; }
         [JsonProperty("type")] public string Type { get; set; }
-        [JsonProperty("revitId")] public int RevitId { get; set; }
+        [JsonProperty("revitId")] public long RevitId { get; set; }
         [JsonProperty("materials")] public List<string> Materials { get; set; }
         [JsonProperty("parameters")] public Dictionary<string, Dictionary<string, object>> Parameters { get; set; }
         [JsonProperty("hostId")] public string HostId { get; set; }
